+++ changelog.md
@@ -10,6 +10,7 @@
 - add bluetooth plugin to dock.
 - add notifications plugin to dock.
 - add system tray plugin to dock.
+- add keyboard selector plugin to dock.
 
 ### enhancements